Cast to '(unsigned long)' to squelch warning from 64-bit Cygwin (in which a 'DWORD' is a 32-bit 'unsigned int').

+++ src/inet_util.c
@@ -489,7 +489,7 @@ static void CALLBACK download_callback (HINTERNET hnd,
         context->bytes_read     = ctx->inet_buf.dwBufferLength;
         context->bytes_written += (DWORD) fwrite (context->file_buf, 1, (size_t)context->bytes_read, context->fil);
         ctx->inet_buf.dwBufferLength = sizeof(context->file_buf);
-        TRACE (1, "  InternetReadFileExA(): rc: %d, %lu bytes\n", rc, context->bytes_read);
+        TRACE (1, "  InternetReadFileExA(): rc: %d, %lu bytes\n", rc, (unsigned long)context->bytes_read);
       }
       while (rc && context->bytes_read > 0);
